Only kill process if we haven't exited yet

+++ src/Cafe/Cafe.Waiter.AcceptanceTests/ExternalProcess.cs
@@ -92,7 +92,11 @@ namespace Cafe.Waiter.AcceptanceTests
         {
             Console.WriteLine($"Stopping process: {_executable} {_arguments}");
             Stop();
-            _process.Kill();
+            if (!_process.HasExited)
+            {
+                _process.Kill();
+            }
+
             _process?.Dispose();
         }
